Change card body design.
