+++ meteor-app/imports/ui/pages/search-by-map/searchResultItemRenderers/index.js
@@ -3,5 +3,5 @@
  * so the component files don't need to worry about name conflicts.
  */
 
-export SEARCH_RESULT_ITEM__DEFAULT from './default';
+export SEARCH_RESULT_ITEM__DATASET from './dataset';
 export SEARCH_RESULT_ITEM__PALEOCARRUN from './paleocar-run';
