+++ meteor-app/imports/ui/styling/pages/search.less
@@ -4,7 +4,7 @@
 }
 
 .page-search {
-  @gapBetweenSearchResultItems: 5px;
+  @gapBetweenSearchResultItems: 10px;
 
   .flex-row;
 
@@ -61,7 +61,6 @@
 
   .search-result-item {
     display: block;
-    height: 108px;
     width: 100%;
     box-sizing: content-box;
     // Leave some padding for the card shadow inside.
@@ -176,6 +175,18 @@
       padding-left: 1.2em;
       margin-top: 1em;
       margin-bottom: 1em;
+    }
+  }
+
+  .search-result-item__description {
+    font-size: 12px;
+    line-height: 125%;
+    & > p:first-child {
+      margin-top: 0;
+    }
+    p {
+      margin-top: 0.5em;
+      margin-bottom: 0.5em;
     }
   }
